Switch rpath to use $ORIGIN instead of .
rpath is a list of directories to search for dependencies. $ORIGIN is a
special constant that represents the directory where the executable
currently is. Using relative paths will be relative to the current
working directory.  It makes more sense to me to search for libs
relative to the binary rather than the CWD.